Making Github auth failure trigger OAuth flow
