+++ util.py
@@ -74,26 +74,6 @@ def get_directory_response(path):
     
     return Response(html, headers={'Content-Type': 'text/html', 'Cache-Control': 'no-store private'})
 
-def get_circle_artifacts(owner, repo, ref, github_token):
-    '''
-    '''
-    import requests, urlparse, os.path
-    
-    circle_token = 'a17131792f4c4bcb97f2f66d9c58258a0ee0e621'
-
-    github_auth = github_token.get('access_token'), 'x-oauth-basic'
-    status_url = 'https://api.github.com/repos/{owner}/{repo}/statuses/{ref}'.format(**locals())
-    status = requests.get(status_url, auth=github_auth).json()[0]
-
-    circle_url = status['target_url'] if (status['state'] == 'success') else None
-    circle_build = os.path.relpath(urlparse.urlparse(circle_url).path, '/gh/')
-
-    artifacts_url = 'https://circleci.com/api/v1/project/{}/artifacts?circle-token={}'.format(circle_build, circle_token)
-    artifacts = {os.path.relpath(a['pretty_path'], '$CIRCLE_ARTIFACTS'): '{}?circle-token={}'.format(a['url'], circle_token)
-                 for a in requests.get(artifacts_url, headers=dict(Accept='application/json')).json()}
-    
-    return artifacts
-
 def errors_logged(route_function):
     '''
     '''
